requirements.txt
# ================================================
# Requirements - Sistema Multi-Agente de Gestión de Contactos
# ================================================

# ========================================
# Core Dependencies
# ========================================
python-telegram-bot==21.0
google-generativeai==0.5.0
httpx==0.27.0
pydantic==2.6.0
pydantic-settings==2.2.0

# ========================================
# Database
# ========================================
SQLAlchemy==2.0.27
psycopg2-binary==2.9.9

# ========================================
# Utilities
# ========================================
python-dotenv==1.0.0
tenacity==8.2.0
structlog==24.1.0

# ========================================
# Testing
# ========================================
pytest==8.0.0
pytest-asyncio==0.23.0
pytest-cov==4.1.0
pytest-mock==3.12.0

# ========================================
# Development
# ========================================
black==24.1.0
flake8==7.0.0
mypy==1.8.0
